+++ Sources/TABResourceLoader/Protocols/Resource types/JSONDictionaryResourceType.swift
@@ -40,11 +40,32 @@ public protocol JSONDictionaryResourceType: DataResourceType {
    - returns: An instantiated model if parsing was successful, otherwise throws
    */
   func model(from jsonDictionary: [String : Any]) throws -> Model
+
+  /**
+   Optionally returns an error from the JSON dictionary provided
+   
+   - parameter jsonDictionary: The JSON dictionary to parse
+   
+   - returns: An error from the jsonDictionary
+   */
+  func error(from jsonDictionary: [String : Any]) -> Error?
 }
 
 extension JSONDictionaryResourceType {
 
   public func model(from data: Data) throws -> Model {
+    let jsonDictionary = try self.jsonDictionary(from: data)
+    return try model(from: jsonDictionary)
+  }
+
+  public func error(from data: Data) -> Error? {
+    guard let jsonDictionary = try? self.jsonDictionary(from: data) else {
+      return nil
+    }
+    return error(from: jsonDictionary)
+  }
+
+  public func jsonDictionary(from data: Data) throws ->  [String : Any] {
     guard let jsonObject = try? JSONSerialization.jsonObject(with: data, options: .mutableContainers) else {
       throw JSONParsingError.invalidJSONData
     }
@@ -53,7 +74,14 @@ extension JSONDictionaryResourceType {
       throw JSONParsingError.notAJSONDictionary
     }
 
-    return try model(from: jsonDictionary)
+    return jsonDictionary
   }
 
+}
+
+// No errors are returned by default
+public extension JSONDictionaryResourceType {
+  func error(from jsonDictionary: [String : Any]) -> Error? {
+    return nil
+  }
 }