+++ Sources/TABResourceLoader/Services/NetworkDataResourceService.swift
@@ -34,13 +34,15 @@ import Foundation
 /// - noHTTPURLResponse:        No HTTPURLResponse exists
 /// - sessionError:             The networking error returned by the URLSession
 /// - statusCodeError:          A status code error between 400 and 600 (not including 600) was returned
-/// - couldNotParseData:        The Data could not be parsed
+/// - noDataProvided:           No data was returned
+/// - statusCodeCustomError:    Contains the custom error for failed status code
 public enum NetworkServiceError: Error {
   case couldNotCreateURLRequest
   case noHTTPURLResponse
   case sessionError(error: Error)
   case statusCodeError(statusCode: Int)
-  case couldNotParseData(error: Error)
+  case noDataProvided
+  case statusCodeCustomError(statusCode: Int, error: Error)
 }
 
 /// Object used to retrive types that conform to both @NetworkResourceType and DataResourceType
@@ -89,8 +91,7 @@ open class NetworkDataResourceService {
   @discardableResult
   func fetch<Resource: NetworkResourceType & DataResourceType>(resource: Resource, networkServiceActivity: NetworkServiceActivity, completion: @escaping (NetworkResponse<Resource.Model>) -> Void) -> Cancellable? {
     guard var urlRequest = resource.urlRequest() else {
-      let parsedResult = Result<Resource.Model>.failure(NetworkResponseHandlerError.noDataProvided)
-      completion(.failure(parsedResult, nil, NetworkServiceError.couldNotCreateURLRequest))
+      completion(.failure(.couldNotCreateURLRequest, nil))
       return nil
     }