+++ Sources/TABResourceLoader/Services/NetworkResponseHandler.swift
@@ -26,9 +26,8 @@ import Foundation
 /// # Failure case
 ///
 /// Defines a response that failed, contains:
-///   - The Result of parsing the model, success case contains the model and failure case contains the parsing error
-///   - A HTTPURLResponse if it exists
 ///   - A NetworkServiceError
+///   - A HTTPURLResponse if it exists
 ///
 /// ## Reasons for failure can be (in order):
 ///   1. No HTTPURLResponse with a URLSession error
@@ -38,51 +37,56 @@ import Foundation
 ///   5. Could not parse the data
 public enum NetworkResponse<Model> {
   case success(Model, HTTPURLResponse)
-  case failure(Result<Model>, HTTPURLResponse?, NetworkServiceError)
-}
-
-enum NetworkResponseHandlerError: Error {
-  case noDataProvided
+  case failure(NetworkServiceError, HTTPURLResponse?)
 }
 
 struct NetworkResponseHandler {
 
   static func resultFrom<Resource: NetworkResourceType & DataResourceType>(resource: Resource, data: Data?, URLResponse: Foundation.URLResponse?, error: Error?) -> NetworkResponse<Resource.Model> {
 
-    let parsedResult = self.parsedResult(transformedFrom: data, using: resource)
-
     guard let HTTPURLResponse = URLResponse as? HTTPURLResponse else {
       if let error = error {
-        return .failure(parsedResult, nil, .sessionError(error: error))
+        return .failure(.sessionError(error: error), nil)
       } else {
-        return .failure(parsedResult, nil, .noHTTPURLResponse)
+        return .failure(.noHTTPURLResponse, nil)
       }
     }
 
+    guard let data = data else {
+      let networkError = self.networkServiceError(HTTPURLResponse: HTTPURLResponse, error: error) ?? NetworkServiceError.noDataProvided
+      return .failure(networkError, HTTPURLResponse)
+    }
+
     switch HTTPURLResponse.statusCode {
-    case 400..<600:
-      return .failure(parsedResult, HTTPURLResponse, .statusCodeError(statusCode: HTTPURLResponse.statusCode))
-    default: break
+    case 200..<400:
+      let parsedResult = Result<Resource.Model> {
+        return try resource.model(from: data)
+      }
+      switch parsedResult {
+      case .success(let model):
+        return .success(model, HTTPURLResponse)
+      case .failure:
+        return .failure(.statusCodeError(statusCode: HTTPURLResponse.statusCode), HTTPURLResponse)
+      }
+    default:
+      if let error = resource.error(from: data) {
+        return .failure(.statusCodeCustomError(statusCode: HTTPURLResponse.statusCode, error: error), HTTPURLResponse)
+      } else {
+        return .failure(.statusCodeError(statusCode: HTTPURLResponse.statusCode), HTTPURLResponse)
+      }
     }
 
-    if let error = error {
-      return .failure(parsedResult, HTTPURLResponse, .sessionError(error: error))
-    }
-
-    switch parsedResult {
-    case .success(let model):
-      return .success(model, HTTPURLResponse)
-    case .failure(let parsingError):
-       return .failure(parsedResult, HTTPURLResponse, .couldNotParseData(error: parsingError))
-    }
   }
 
-  private static func parsedResult<Resource: DataResourceType>(transformedFrom data: Data?, using resource: Resource) -> Result<Resource.Model> {
-    guard let data = data else {
-      return .failure(NetworkResponseHandlerError.noDataProvided)
-    }
-    return Result<Resource.Model> {
-      return try resource.model(from: data)
+  private static func networkServiceError(HTTPURLResponse: HTTPURLResponse, error: Error?) -> NetworkServiceError? {
+    switch HTTPURLResponse.statusCode {
+    case 400..<600:
+      return .statusCodeError(statusCode: HTTPURLResponse.statusCode)
+    default:
+      if let error = error {
+        return .sessionError(error: error)
+      }
+      return nil
     }
   }