+++ Tests/TABResourceLoaderTests/Services/NetworkDataResourceServiceTests.swift
@@ -76,7 +76,7 @@ class NetworkDataResourceServiceTests: XCTestCase {
         switch result {
         case .success:
           XCTFail("No error found")
-        case .failure(_, _, let error):
+        case .failure(let error, _):
           guard case NetworkServiceError.couldNotCreateURLRequest = error else {
             XCTFail("Unexpected error: \(error)")
             return
@@ -104,9 +104,9 @@ class NetworkDataResourceServiceTests: XCTestCase {
         switch result {
         case .success:
           XCTFail("No error found")
-        case .failure(_, _, let error):
+        case .failure(let error, _):
           guard case NetworkServiceError.statusCodeError(let statusCode) = error else {
-            XCTFail("Unexpected error: \(error)")
+            XCTFail("Unexpected error: \(error)", file: file, line: lineNumber)
             return
           }
           XCTAssertEqual(statusCode, expectedStatusCode)
@@ -117,18 +117,19 @@ class NetworkDataResourceServiceTests: XCTestCase {
     }
   }
 
-  func test_fetch_whenSessionCompletesDataAndHTTPURLResponse_callsSuccess() {
+  func test_fetch_whenSessionCompletesDataAndHTTPURLResponse_callsSuccessWhenParsingSucceed() {
     performAsyncTest { expectation in
       testService.fetch(resource: mockResource) { result in
         switch result {
         case .success(let model, _):
           XCTAssertEqual(model, "")
           expectation?.fulfill()
-        case .failure(_, _, let error):
+        case .failure(let error, _):
           XCTFail("Unexpected error: \(error)")
         }
       }
-      mockSession.capturedCompletion!(Data(), HTTPURLResponse(), nil)
+      let successResponse = HTTPURLResponse(url: mockURL, statusCode: 200, httpVersion: nil, headerFields: nil)
+      mockSession.capturedCompletion!(Data(), successResponse, nil)
     }
   }
 
@@ -147,9 +148,9 @@ class NetworkDataResourceServiceTests: XCTestCase {
         switch result {
         case .success:
           XCTFail("No error found")
-        case .failure(_, _, let error):
+        case .failure(let error, _):
           guard case NetworkServiceError.sessionError(let testError) = error else {
-            XCTFail("Unexpected error: \(error)")
+            XCTFail("Unexpected error: \(error)", file: file, line: lineNumber)
             return
           }
           XCTAssertEqual(testError._domain, expectedError._domain)
@@ -168,7 +169,7 @@ class NetworkDataResourceServiceTests: XCTestCase {
         switch result {
         case .success:
           XCTFail("No error found")
-        case .failure(_, _, let error):
+        case .failure(let error, _):
           guard case NetworkServiceError.sessionError(let testError) = error else {
             XCTFail("Unexpected error: \(error)")
             return
@@ -187,8 +188,8 @@ class NetworkDataResourceServiceTests: XCTestCase {
         switch result {
         case .success:
           XCTFail("No error found")
-        case .failure(_, _, let error):
-          guard case NetworkServiceError.couldNotParseData(error: NetworkResponseHandlerError.noDataProvided) = error else {
+        case .failure(let error, _):
+          guard case NetworkServiceError.noDataProvided = error else {
             XCTFail("Unexpected error: \(error)")
             return
           }